// src/app/admin/models/orden-servicio.ts
import { Agencia } from "./agencia";
import { AgenciaDestino } from "./agencia-destino";
import { ClienteProveedor } from "./cliente-proveedor";
import { TipoEnvio } from "./tipo-envio";
import { TipoOrdenServicio } from "./tipo-orden-servicio";
import { TipoPago } from "./tipo-pago";
import { TipoServicio } from "./tipo-servicio";
import { Ubigeo } from "./ubigeo";
import { TipoEmbalaje } from './tipo-embalaje';

export class OrdenServicio{
  idOrdenServicio: string;
  agenciaOrigen: Agencia;
  origen: Ubigeo;
  agenciaDestino: Agencia;
  destino: Ubigeo;
  cliente: ClienteProveedor;
  area: string = '';
  remitente: string = '';
  consignatario: string = '';
  idConsignatario: string = '';
  direccion: string = '';
  telefono: string = '';
  email: string = '';
  guia: string = '';
  referencia1: string = '';
  referencia2: string = '';
  referencia3: string = '';
  contenido: string = '';
  piezas: number;
  peso: number;
  pesoMinimo: number;
  tarifaMinima: number;
  tarifaExeso: number;
  adicionalPod: number;
  adicionalTransferencia: number;
  adicionalOtro: number;
  subtotal: number;
  igv: number;
  total: number;
  tipoOrdenServicio: string;
  tipoPago: string;
  tipoEnvio: string;
  tipoServicio: string;
  tipoEmbalaje: string;
  idManifiestoSalida: number;
  idManifiestoReparto: number;
  idFacturacion: number;
  idLiquidacionCliente: number;
  idLiquidacionAgencia: number;
  fechaHoraRegistro: string;
  fechaHoraManifiestoSalida: string;
  fechaHoraDespacho: string;
  fechaHoraLlegada: string;
  fechaHoraReparto: string;
  fechaHoraEntrega: string;
  recepcionista: string;
  dniRecepcionista: string;
  podVirtual: string;
  estadoOperacion: string;
  estadoFacturacion: string;
}